+++ stroom-index/stroom-index-impl-db/src/main/java/stroom/index/impl/db/IndexVolumeDaoImpl.java
@@ -176,7 +176,7 @@ class IndexVolumeDaoImpl implements IndexVolumeDao {
         int numberOfRows = 1000000;
 
         if (pageRequest != null) {
-            offset = pageRequest.getOffset().intValue();
+            offset = pageRequest.getOffset();
             numberOfRows = pageRequest.getLength();
         }
